Added Grouping of Units
Allow for units to be grouped and added automated software thermostat.   In this way, a number of units can be defined as a group.   Some state variables (temp, humidity) are computed for the group as an average of the stat of each unit in the group.    Optionally, an automated control function can be activated which overrides the thermostatic control each unit has.   The automated function will monitor the temperature of the group and turn the units on and off in unison in response to the temperature of the room.

Also fixed a bug so that temperature is reported correctly so that HomeKit and Indigo can report the termperature of any unit.